Further work on the LDAP connection pool.
It more or less works now, except for a small annoyance when reviving
connections from the dead list. Currently when one dead server has been
revived, the loop exits and leaves the remaining dead servers in the list.

This has some clear benefits if the dead list is very long, but it is perhaps
a bit annoying when the list is short.