Call_Queue package is more or less done.
It's perhaps still a bit rough around the edges, but it gets the job done.

It would perhaps be nice with an interface where the client can request a queue
with specific priority or perhaps limit the length of the returned queue (implies
sorting the calls serverside, so only the oldest calls are returned).

Lots of interesting stuff can be done. Currently we simply support returning the entire
queue.