Added the /get/call interface.
This interface can either be called as is, or with the ?id=n parameter. if
called with the id parameter, we search for a call with the given id in
the queue, and if found it is removed from the queue and the id is returned
in a JSON object. If the id does not exist in the queue, and empty JSON
object is returned.

A call to /get/call without the id parameter will simply return the next
available call in the queue. If a call is found, a JSON object is returned
containing the id of the call. If no calls are found, an empty JSON object
is returned.