Change Call_Queue handling Structure.
Before the Call_Queue_JSON made the call to the Call_Queue, took that
response and made it into JSON, and last wrap it in a reponse object.
All of this is nothing to do with turning it into a JSON object.
Therefor i made a new handler package, to call Call_Queue about the
information, send it to Call_Queue_JSON, so the information is in JSON
format, and last it takes that and send it out. So now the packages have
a higher cohesion.