Changed Event_Parser types. Fixed a bug in Call_Queue_JSON
Call_Queue_JSON had a bug in Convert_Queue, where the cursor
is never moved forward.

Event_Parser parsed an event out to an Ordered_Map, which
could mislead somebody into thinking that the order of
the data matters, which it don't. So now it is parsed
into a Hashed_Map.

All the keywords like "Event, Channel and Uniqueid" are
now an enum type, that are used as key_type in the Hashed_Map.