Completed the System_Message package.
Instead of rampant use of the Yolk.Log package we can centralize our
critical/error/info system messages for tighter conrol over what is written
where and how.

The end goal of this package is to more or less completely remove the use of
Yolk.Log in Alice, except perhaps for debug messages.