Added support for SQLite.
This means some breaking changes since we've dumped the database connector.
Alice is no longer aware of more than one database. There is no fallback. If
fallback is required in a setup, then this is something that must be done outside
of Alice, ie. using database connection proxies or failover systems. Also this
change means that Alice now defaults to a local SQLite database. It should be as
easy as possible to get her up and running, and a local SQLite database is in
every possible way easier to get going than a PostgreSQL (or other proper RDBMS).